+++ src/Exceptions/InvalidMessageType.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * @author hollodotme
+ */
+
+namespace hollodotme\FluidValidator\Exceptions;
+
+/**
+ * Class InvalidMessageType
+ * @package hollodotme\FluidValidator\Exceptions
+ */
+final class InvalidMessageType extends FluidValidatorException
+{
+
+}